Track1/8_冬瓜皮/code/requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
auto_gptq==0.7.1+cu118
-e git+https://github.com/intel/auto-round.git@09f1ab49127c1b0bc000a22ecd776c6f0c8acea5#egg=auto_round
autoawq==0.2.8
bitsandbytes==0.45.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
DataProperty==1.1.0
datasets==3.2.0
Deprecated==1.2.15
dill==0.3.8
diskcache==5.6.3
embedding_quant_cuda==0.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gekko==1.2.1
gguf==0.14.0
h11==0.14.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.7
llama_cpp_python==0.3.7
llvmlite==0.43.0
-e git+https://github.com/EleutherAI/lm-evaluation-harness@888ac292c5ef041bcae084e7141e50e154e1108a#egg=lm_eval
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mbstrdecoder==1.1.3
modelscope==1.22.0
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
neural_compressor==3.2
nltk==3.9.1
numba==0.60.0
numexpr==2.10.2
numpy==2.0.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
opencv-python-headless==4.11.0.86
optimum==1.23.3
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
peft==0.14.0
pillow==11.1.0
portalocker==3.1.1
prettytable==3.12.0
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pybind11==2.13.6
pycocotools==2.0.8
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyparsing==3.2.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rouge==1.0.1
rouge_score==0.1.2
sacrebleu==2.5.1
safetensors==0.5.0
schema==0.7.7
scikit-learn==1.6.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sqlitedict==2.1.0
sse-starlette==2.2.1
starlette==0.45.3
starlette-context==0.3.6
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tbb==2022.0.0
tcmlib==1.2.0
tcolorpy==0.1.7
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.47.1
triton==2.3.1
typepy==1.3.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
word2number==1.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
